+++ CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to the "kleio" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [1.0.18] - 2024-12-11
+## [1.0.18] - 2024-12-12
 
 ### Added
 - Added option to delete all generated files in a selected folder.
 
 ### Fixed
 - Fixed Docker port binding issues when starting multiple containers.
+- Fixed issue of looping error handling when Docker is not running.
 
 ## [1.0.17] - 2024-11-29